Fixing a bug that allowed users to get free funds
If a user were to have a negative balance, attempt to make a journey,
be informed of their negative balance, then enter a destination station,
they would be 'refunded' the difference between the maximum charge and
the fare between the two entered stations, despite never having been
charged.
This resulted in a free top-up for the user, but it has now been
patched.